extract layout from app, dialog from workspace, other refactors
